refactor(dapp): TokenView and TokenCard components
