+++ packages/dapp/src/pages/MyTokens.tsx
@@ -1,34 +1,17 @@
-import type { StayToken, TokenData } from 'stays-core';
-import { ReactChild, useCallback, useMemo, useState } from 'react';
-import * as Icons from 'grommet-icons';
-import { Grid, Spinner, Button, Box, Image, Text } from 'grommet';
+import type { TokenData } from 'stays-core';
+import { useCallback, useMemo } from 'react';
+import { Spinner, Box, Text } from 'grommet';
 import { PageWrapper } from './PageWrapper';
 import { MessageBox } from '../components/MessageBox';
-import { CustomText, StayVoucherQr } from '../components/StayVoucherQr';
-import { useSearchParams, useNavigate } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 import { useAppState } from '../store';
 // import { useWindowsDimension } from "../hooks/useWindowsDimension";
 import { useMyTokens, useGetToken } from '../hooks/useMyTokens';
-import { useContract } from '../hooks/useContract';
-import { centerEllipsis } from '../utils/strings';
-import { getNetwork } from '../config';
-import { ExternalLink } from '../components/ExternalLink';
 // import { useGoToMessage } from '../hooks/useGoToMessage';
-import { LodgingFacilityRecord } from '../store/actions';
 import styled from 'styled-components';
-import { utils, BigNumber as BN } from 'ethers';
+import { TokenCard } from '../components/tokens/TokenCard';
+import { TokenView } from '../components/tokens/TokenView';
 // import { CustomButton } from '../components/SearchResultCard';
-import { getDate } from '../utils/dates';
-
-const HotelTitle = styled(Text)`
-  color: #000;
-  font-weight: 900;
-  font-size: 22px;
-  font-family: 'Inter';
-  line-height: 28px;
-  margin-bottom: .5rem;
-`;
-
 
 export const Header = styled(Text)`
   font-weight: 900;
@@ -38,16 +21,6 @@ export const Header = styled(Text)`
   color: #0D0E0F;
   margin-bottom: 2rem;
   margin-top: 2rem;
-`;
-
-export const CustomBoldText = styled(Text)`
-  color: #0D0E0F;
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 600;
-  font-size: 18px;
-  line-height: 24px;
-  // margin-bottom: 1.5rem;
 `;
 
 // const ResponsiveColumn = (winWidth: number): string[] => {
@@ -66,272 +39,6 @@ export const CustomBoldText = styled(Text)`
 //   }
 //   return [];
 // };
-
-export interface TokenCardProps extends TokenData {
-  onClick?: () => void,
-  children?: ReactChild | null,
-  facility?: LodgingFacilityRecord;
-}
-
-export interface TokenViewProps extends StayToken {
-  facilityOwner: string | undefined;
-  facility: LodgingFacilityRecord | undefined;
-  withCloseButton?: boolean;
-  withRpcProvider?: boolean;
-}
-
-export const TokenCard = ({
-  image,
-  name,
-  description,
-  attributes,
-  onClick = () => { },
-  facility,
-  children
-}: TokenCardProps) => {
-  console.log('facility', facility)
-  if (!facility || !attributes) {
-    return null
-  }
-  const parseTrait = (trait: string): any => {
-    return (attributes || []).find(attr => attr.trait_type === trait)?.value ?? ''
-  };
-  const space = facility.spaces.find(space => space.contractData.spaceId === parseTrait('spaceId').toLowerCase())
-  const quantity = Number(parseTrait('quantity'))
-  const numberOfDays = Number(parseTrait('numberOfDays'))
-  const total = BN.from(space?.contractData.pricePerNightWei || 0).mul(BN.from(numberOfDays)).mul(BN.from(quantity)).toString();
-  const totalEther = utils.formatUnits(total, 'ether');
-
-  return (
-    <Box>
-      <Box
-        direction='row'
-        justify='between'
-        pad='small'
-        style={{ borderBottom: '1px solid black' }}
-        onClick={() => onClick()}
-      >
-        <Box margin={{ right: 'large' }}>
-          <Image
-            height='120'
-            width='120'
-            fit='cover'
-            src={image}
-          />
-        </Box>
-        <Box pad='small'>
-          <HotelTitle>{facility.name}</HotelTitle>
-          <CustomText>{facility.address.streetAddress}, {facility.address.postalCode} {facility.address.locality}, {facility.address.country}. </CustomText>
-          <CustomText>{space?.name},{quantity} {quantity === 1 ? 'room' : 'rooms'} </CustomText>
-        </Box>
-        <Box align='center' justify='center' pad='small'>
-          <CustomText>{getDate(parseTrait('startDay')).toFormat('dd.MM.yyyy')} - {getDate(Number(parseTrait('startDay')) + Number(parseTrait('numberOfDays'))).toFormat('dd.MM.yyyy')}</CustomText>
-        </Box>
-        <Box
-          alignSelf='center'
-          style={{ justifySelf: 'end' }}
-          pad={{ horizontal: 'small' }}
-        >
-          <Box pad='small'>
-            <Text size="xlarge" weight="bold">
-              {totalEther}
-            </Text>
-          </Box>
-        </Box>
-      </Box>
-      {children}
-    </Box>
-  );
-};
-
-export const TokenView = ({
-  tokenId,
-  owner,
-  facilityOwner,
-  facility,
-  status,
-  data: {
-    name,
-    description,
-    image,
-    attributes
-  },
-  withCloseButton = true,
-  withRpcProvider = false
-}: TokenViewProps) => {
-  const { provider, rpcProvider, ipfsNode } = useAppState();
-  const [, , contractError] = useContract(
-    withRpcProvider ? rpcProvider : provider,
-    ipfsNode
-  );
-  const navigate = useNavigate();
-  // const showMessage = useGoToMessage();
-  // const [cancelLoading, setCancelLoading] = useState<boolean>(false);
-  const [cancellationTxHash,] = useState<string | undefined>();
-  const [error, setError] = useState<string | undefined>();
-  const cancellationTxHashLink = useMemo(() => {
-    const network = getNetwork()
-    return cancellationTxHash ? `${network.blockExplorer}/tx/${cancellationTxHash}` : '#'
-  }, [cancellationTxHash]);
-
-  // const cancelTx = useCallback(
-  //   async () => {
-  //     setError(undefined);
-  //     setCancellationTxHash(undefined);
-  //     if (!contract) {
-  //       return;
-  //     }
-  //     try {
-  //       setCancelLoading(true);
-  //       await contract.cancel(tokenId, undefined, setCancellationTxHash);
-  //       setCancelLoading(false);
-  //       // Show success message;
-  //       showMessage(
-  //         `Stay token #${tokenId} is successfully cancelled. All funds are fully refunded`,
-  //         'info'
-  //       );
-  //     } catch (err) {
-  //       setError((err as Error).message || 'Unknown cancellation error');
-  //       setCancelLoading(false);
-  //     }
-  //   },
-  //   [showMessage, contract, tokenId]
-  // );
-
-  if (!attributes || !facility) {
-    return null;
-  }
-
-  return (
-    <Box
-      alignSelf='center'
-      direction='column'
-      pad={{ bottom: 'medium' }}
-      style={{ position: 'relative' }}
-    >
-      <Box
-        direction='row'
-        justify='between'
-        pad='large'
-        round={false}
-        width='65rem'
-        style={{ borderBottom: '1px solid black' }}
-      >
-        <Box>
-          <Grid
-            fill='horizontal'
-            pad='small'
-            columns={['small', 'auto']}
-            responsive
-          >
-            <Box>
-              <CustomBoldText weight='bold'>Status</CustomBoldText>
-            </Box>
-            <Box>
-              <CustomText>{status ?? 'unknown'}</CustomText>
-            </Box>
-          </Grid>
-          <Grid
-            fill='horizontal'
-            pad='small'
-            columns={['small', 'auto']}
-            responsive
-          >
-            <Box>
-              <CustomBoldText weight='bold'>Name</CustomBoldText>
-            </Box>
-            <Box>
-              <CustomText>{name}</CustomText>
-            </Box>
-          </Grid>
-          <Grid
-            fill='horizontal'
-            pad='small'
-            columns={['small', 'auto']}
-            responsive
-          >
-            <Box>
-              <CustomBoldText weight='bold'>Description</CustomBoldText>
-            </Box>
-            <Box>
-              <CustomText>{description}</CustomText>
-            </Box>
-          </Grid>
-        </Box>
-
-        <Box pad={{ vertical: 'small', horizontal: 'large' }} justify='between'>
-          {!withRpcProvider &&
-            <StayVoucherQr
-              provider={provider}
-              from={owner}
-              to={facilityOwner}
-              tokenId={tokenId}
-              onError={err => setError(err)}
-              name={name}
-              description={description}
-              attributes={attributes}
-              facility={facility}
-              pricePerNightWei={'0'}
-            />
-          }
-
-          {status === 'booked' &&
-            <Box>
-              {/* <CustomButton
-                label={
-                  <Box direction='row'>
-                    <Box>
-                      <Text>Cancel and get refund</Text>
-                    </Box>
-                    {cancelLoading &&
-                      <Box pad={{ left: 'small' }}>
-                        <Spinner />
-                      </Box>
-                    }
-                  </Box>
-                }
-                disabled={cancelLoading}
-                onClick={cancelTx}
-              /> */}
-              {!!cancellationTxHash
-                ? <ExternalLink
-                  href={cancellationTxHashLink}
-                  label={centerEllipsis(cancellationTxHash)}
-                />
-                : null
-              }
-            </Box>
-          }
-          {withCloseButton &&
-            <Button
-              style={{
-                position: 'absolute',
-                top: 10,
-                right: 10
-              }}
-
-              icon={<Icons.Close color="plain" />}
-              hoverIndicator
-              onClick={() => navigate('/tokens', { replace: true })}
-            />
-          }
-        </Box>
-      </Box>
-
-      <MessageBox type='error' show={!!error}>
-        <Box>
-          {error}
-        </Box>
-      </MessageBox>
-
-      <MessageBox type='error' show={!!contractError}>
-        <Box>
-          {contractError}
-        </Box>
-      </MessageBox>
-    </Box >
-  );
-};
 
 export const MyTokens = () => {
   const { provider, ipfsNode, account, lodgingFacilities } = useAppState();
